Test login

+++ BibliothequeAdministration/src/main/java/com/ocp7bibliotheque/bibliothequeadministration/DAO/RoleRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface RoleRepository extends JpaRepository<Role,Integer> {
 
-    public Optional<Role> findById(long id);
+    public Optional<Role> findById(int id);
     public Optional<Role> findByName(String name);
 
 }
